docs: Update account's vaultCipher relationships format
In https://github.com/cozy/cozy-doctypes/pull/119 we wrongly introduced
an array in io.cozy.account's vaultCipher relationships. This could let
somebody think that an account can have a relationship to multiple
ciphers, which is not the case. Since an account can only be linked to
one cipher, the value of `relationships.vaultCipher.data` should be an
object.